requirements.txt
# Data Collection
requests>=2.31.0
beautifulsoup4>=4.12.0
python-dateutil>=2.8.2
lxml>=4.9.0

# Data Processing
pandas>=2.1.0
numpy>=1.24.0

# Machine Learning
scikit-learn>=1.3.0
catboost>=1.2.0
xgboost>=2.0.0

# Explainability
shap>=0.44.0
lime>=0.2.0.1

# Visualization
matplotlib>=3.7.0
seaborn>=0.13.0

# Utilities
tqdm>=4.66.0
jupyter>=1.0.0
notebook>=7.0.0
